@@ -1,13 +1,14 @@
-// swift-tools-version:5.7
+// swift-tools-version:5.9
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
 
 let package = Package(
     name: "hummingbird",
-    platforms: [.macOS(.v13), .iOS(.v16), .tvOS(.v16)],
+    platforms: [.macOS(.v14), .iOS(.v17), .tvOS(.v17)],
     products: [
         .library(name: "Hummingbird", targets: ["Hummingbird"]),
+        .library(name: "HummingbirdCore", targets: ["HummingbirdCore"]),
         .library(name: "HummingbirdFoundation", targets: ["HummingbirdFoundation"]),
         .library(name: "HummingbirdJobs", targets: ["HummingbirdJobs"]),
         .library(name: "HummingbirdXCT", targets: ["HummingbirdXCT"]),
@@ -18,11 +19,11 @@ let package = Package(
         .package(url: "https://github.com/apple/swift-log.git", from: "1.4.0"),
         .package(url: "https://github.com/apple/swift-metrics.git", "1.0.0"..<"3.0.0"),
         .package(url: "https://github.com/apple/swift-distributed-tracing.git", from: "1.0.1"),
-        .package(url: "https://github.com/apple/swift-nio.git", from: "2.60.0"),
+        .package(url: "https://github.com/apple/swift-nio.git", from: "2.62.0"),
         .package(url: "https://github.com/apple/swift-nio-extras.git", from: "1.7.0"),
         .package(url: "https://github.com/apple/swift-nio-http2.git", from: "1.28.0"),
         .package(url: "https://github.com/apple/swift-nio-ssl.git", from: "2.14.0"),
-        .package(url: "https://github.com/apple/swift-nio-transport-services.git", from: "1.18.0"),
+        .package(url: "https://github.com/apple/swift-nio-transport-services.git", from: "1.20.0"),
         .package(url: "https://github.com/swift-server/swift-service-lifecycle.git", from: "2.0.0"),
     ],
     targets: [
@@ -56,6 +57,7 @@ let package = Package(
             .product(name: "NIOHTTP1", package: "swift-nio"),
         ]),
         .target(name: "HummingbirdCore", dependencies: [
+            .product(name: "AsyncAlgorithms", package: "swift-async-algorithms"),
             .product(name: "Logging", package: "swift-log"),
             .product(name: "NIOCore", package: "swift-nio"),
             .product(name: "NIOConcurrencyHelpers", package: "swift-nio"),
@@ -72,12 +74,12 @@ let package = Package(
             .product(name: "NIOPosix", package: "swift-nio"),
             .product(name: "NIOSSL", package: "swift-nio-ssl"),
         ]),
-        .target(name: "HummingbirdHTTP2", dependencies: [
-            .byName(name: "HummingbirdCore"),
-            .product(name: "NIOCore", package: "swift-nio"),
-            .product(name: "NIOHTTP2", package: "swift-nio-http2"),
-            .product(name: "NIOSSL", package: "swift-nio-ssl"),
-        ]),
+        /*        .target(name: "HummingbirdHTTP2", dependencies: [
+                .byName(name: "HummingbirdCore"),
+                .product(name: "NIOCore", package: "swift-nio"),
+                .product(name: "NIOHTTP2", package: "swift-nio-http2"),
+                .product(name: "NIOSSL", package: "swift-nio-ssl"),
+            ]),*/
         .target(name: "HummingbirdTLS", dependencies: [
             .byName(name: "HummingbirdCore"),
             .product(name: "NIOCore", package: "swift-nio"),
