@@ -2,7 +2,7 @@
 //
 // This source file is part of the Hummingbird server framework project
 //
-// Copyright (c) 2021-2021 the Hummingbird authors
+// Copyright (c) 2023 the Hummingbird authors
 // Licensed under Apache License v2.0
 //
 // See LICENSE.txt for license information
@@ -12,103 +12,79 @@
 //
 //===----------------------------------------------------------------------===//
 
+import AsyncAlgorithms
 import NIOCore
+import NIOHTTP1
 
-/// Request Body. Either a ByteBuffer or a ByteBuffer streamer
-public enum HBRequestBody: Sendable {
-    /// Static ByteBuffer
-    case byteBuffer(ByteBuffer?)
-    /// ByteBuffer streamer
-    case stream(HBByteBufferStreamer)
+public enum HBRequestBody: Sendable, AsyncSequence {
+    case byteBuffer(ByteBuffer)
+    case stream(HBStreamedRequestBody)
 
-    /// Return as ByteBuffer
-    public var buffer: ByteBuffer? {
+    public typealias Element = ByteBuffer
+    public typealias AsyncIterator = HBStreamedRequestBody.AsyncIterator
+
+    public func makeAsyncIterator() -> HBStreamedRequestBody.AsyncIterator {
         switch self {
-        case .byteBuffer(let buffer):
-            return buffer
-        default:
-            preconditionFailure("Cannot get buffer on streaming RequestBody")
+        case .byteBuffer:
+            /// The server always creates the HBRequestBody as a stream. If it is converted
+            /// into a single ByteBuffer it cannot be treated as a stream after that
+            preconditionFailure("Cannot convert collapsed request body back into a sequence")
+        case .stream(let streamer):
+            return streamer.makeAsyncIterator()
         }
     }
 
-    /// Return as streamer if it is a streamer
-    public var stream: HBStreamerProtocol? {
+    /// Return new HBRequestBody as a single ByteBuffer
+    public func collate(maxSize: Int) async throws -> HBRequestBody {
         switch self {
+        case .byteBuffer:
+            return self
         case .stream(let streamer):
-            return streamer
-        case .byteBuffer(let buffer):
-            guard let buffer = buffer else {
-                return nil
-            }
-            return HBStaticStreamer(buffer)
-        }
-    }
-
-    /// Provide body as a single ByteBuffer
-    /// - Parameter eventLoop: EventLoop to use
-    /// - Returns: EventLoopFuture that will be fulfilled with ByteBuffer. If no body is include then return `nil`
-    @available(*, deprecated, message: "Use the version of `consumeBody` which sets a maximum size for the resultant ByteBuffer")
-    public func consumeBody(on eventLoop: EventLoop) -> EventLoopFuture<ByteBuffer?> {
-        switch self {
-        case .byteBuffer(let buffer):
-            return eventLoop.makeSucceededFuture(buffer)
-        case .stream(let streamer):
-            return streamer.collate(maxSize: .max).hop(to: eventLoop)
-        }
-    }
-
-    /// Provide body as a single ByteBuffer
-    /// - Parameters
-    ///   - maxSize: Maximum size of ByteBuffer to generate
-    ///   - eventLoop: EventLoop to use
-    /// - Returns: EventLoopFuture that will be fulfilled with ByteBuffer. If no body is include then return `nil`
-    public func consumeBody(maxSize: Int, on eventLoop: EventLoop) -> EventLoopFuture<ByteBuffer?> {
-        switch self {
-        case .byteBuffer(let buffer):
-            return eventLoop.makeSucceededFuture(buffer)
-        case .stream(let streamer):
-            return streamer.collate(maxSize: maxSize).hop(to: eventLoop)
+            return try .byteBuffer(await streamer.collect(upTo: maxSize))
         }
     }
 }
 
-extension HBRequestBody: CustomStringConvertible {
-    public var description: String {
-        let maxOutput = 256
-        switch self {
-        case .byteBuffer(let buffer):
-            guard var buffer2 = buffer else { return "empty" }
-            if let string = buffer2.readString(length: min(maxOutput, buffer2.readableBytes)),
-               string.allSatisfy(\.isASCII)
-            {
-                if buffer2.readableBytes > 0 {
-                    return "\"\(string)...\""
-                } else {
-                    return "\"\(string)\""
-                }
-            } else {
-                return "\(buffer!.readableBytes) bytes"
-            }
+/// A type that represents an HTTP request body.
+public struct HBStreamedRequestBody: Sendable, AsyncSequence {
+    public typealias Element = ByteBuffer
 
-        case .stream:
-            return "byte stream"
+    public struct AsyncIterator: AsyncIteratorProtocol {
+        public typealias Element = ByteBuffer
+
+        fileprivate var underlyingIterator: AsyncThrowingChannel<ByteBuffer, Error>.AsyncIterator
+
+        public mutating func next() async throws -> ByteBuffer? {
+            try await self.underlyingIterator.next()
         }
+    }
+
+    /// HBRequestBody is internally represented by AsyncThrowingChannel
+    private var channel: AsyncThrowingChannel<ByteBuffer, Error>
+
+    /// Creates a new HTTP request body
+    @_spi(HBXCT) public init() {
+        self.channel = .init()
+    }
+
+    public func makeAsyncIterator() -> AsyncIterator {
+        AsyncIterator(underlyingIterator: self.channel.makeAsyncIterator())
     }
 }
 
-@available(macOS 10.15, iOS 13.0, tvOS 13.0, watchOS 6.0, *)
-extension HBRequestBody {
-    /// Provide body as a single ByteBuffer
-    /// - Parameters
-    ///   - maxSize: Maximum size of ByteBuffer to generate
-    ///   - eventLoop: EventLoop to use
-    /// - Returns: EventLoopFuture that will be fulfilled with ByteBuffer. If no body is include then return `nil`
-    public func consumeBody(maxSize: Int) async throws -> ByteBuffer? {
-        switch self {
-        case .byteBuffer(let buffer):
-            return buffer
-        case .stream(let streamer):
-            return try await streamer.collate(maxSize: maxSize).get()
-        }
+extension HBStreamedRequestBody {
+    /// push a single ByteBuffer to the HTTP request body stream
+    @_spi(HBXCT) public func send(_ buffer: ByteBuffer) async {
+        await self.channel.send(buffer)
+    }
+
+    /// pass error to HTTP request body
+    @_spi(HBXCT) public func fail(_ error: Error) {
+        self.channel.fail(error)
+    }
+
+    /// Finish HTTP request body stream
+    @_spi(HBXCT) public func finish() {
+        self.channel.finish()
     }
 }
