@@ -21,25 +21,6 @@ import Network
 
 // MARK: Configuration
 
-/// Idle state handlder configuration.
-///
-/// Timeout values before a connection is closed. The `readTimeout`` will trigger after the server
-/// receives no request parts for the specified time. It is meant to catch when a connection is
-/// idle for too long in the middle of reading a request. The writeTimeout will trigger after
-/// the server hasn't written any http parts for the specified time. It is meant to catch if a
-/// connection is idle for too long between requests.
-public struct IdleStateHandlerConfiguration: Sendable {
-    /// timeout when reading a request
-    let readTimeout: TimeAmount
-    /// timeout since last writing a response
-    let writeTimeout: TimeAmount
-
-    public init(readTimeout: TimeAmount = .seconds(30), writeTimeout: TimeAmount = .minutes(3)) {
-        self.readTimeout = readTimeout
-        self.writeTimeout = writeTimeout
-    }
-}
-
 /// Application configuration
 public struct HBApplicationConfiguration: Sendable {
     // MARK: Member variables
@@ -61,12 +42,6 @@ public struct HBApplicationConfiguration: Sendable {
     public let backlog: Int
     /// Allows socket to be bound to an address that is already in use.
     public let reuseAddress: Bool
-    /// Disables the Nagle algorithm for send coalescing.
-    public let tcpNoDelay: Bool
-    /// Pipelining ensures that only one http request is processed at one time
-    public let enableHttpPipelining: Bool
-    /// Idle state handler setup.
-    public let idleTimeoutConfiguration: IdleStateHandlerConfiguration?
     #if canImport(Network)
     /// TLS options for NIO Transport services
     public let tlsOptions: TSTLSOptions
@@ -92,8 +67,6 @@ public struct HBApplicationConfiguration: Sendable {
     ///   - backlog: the maximum length for the queue of pending connections.  If a connection request arrives with the queue full,
     ///         the client may receive an error with an indication of ECONNREFUSE
     ///   - reuseAddress: Allows socket to be bound to an address that is already in use.
-    ///   - tcpNoDelay: Disables the Nagle algorithm for send coalescing.
-    ///   - enableHttpPipelining: Pipelining ensures that only one http request is processed at one time
     ///   - threadPoolSize: Number of threads in application thread pool
     ///   - logLevel: Logging level
     ///   - noHTTPServer: Don't start up the HTTP server.
@@ -105,9 +78,6 @@ public struct HBApplicationConfiguration: Sendable {
         maxStreamingBufferSize: Int = 1 * 1024 * 1024,
         backlog: Int = 256,
         reuseAddress: Bool = true,
-        tcpNoDelay: Bool = false,
-        enableHttpPipelining: Bool = true,
-        idleTimeoutConfiguration: IdleStateHandlerConfiguration? = nil,
         threadPoolSize: Int = 2,
         logLevel: Logger.Level? = nil,
         noHTTPServer: Bool = false
@@ -121,9 +91,6 @@ public struct HBApplicationConfiguration: Sendable {
         self.maxStreamingBufferSize = maxStreamingBufferSize
         self.backlog = backlog
         self.reuseAddress = reuseAddress
-        self.tcpNoDelay = tcpNoDelay
-        self.enableHttpPipelining = enableHttpPipelining
-        self.idleTimeoutConfiguration = idleTimeoutConfiguration
         #if canImport(Network)
         self.tlsOptions = .none
         #endif
@@ -149,12 +116,10 @@ public struct HBApplicationConfiguration: Sendable {
     ///   - maxUploadSize: Maximum upload size allowed for routes that don't stream the request payload
     ///   - maxStreamingBufferSize: Maximum size of data in flight while streaming request payloads before back pressure is applied.
     ///   - reuseAddress: Allows socket to be bound to an address that is already in use.
-    ///   - enableHttpPipelining: Pipelining ensures that only one http request is processed at one time
     ///   - threadPoolSize: Number of threads in application thread pool
     ///   - logLevel: Logging level
     ///   - noHTTPServer: Don't start up the HTTP server.
     ///   - tlsOptions: TLS options for when you are using NIOTransportServices
-    @available(macOS 10.14, iOS 12, tvOS 12, *)
     public init(
         address: HBBindAddress = .hostname(),
         serverName: String? = nil,
@@ -162,8 +127,6 @@ public struct HBApplicationConfiguration: Sendable {
         maxStreamedUploadSize: Int = 4 * 1024 * 1024,
         maxStreamingBufferSize: Int = 1 * 1024 * 1024,
         reuseAddress: Bool = true,
-        enableHttpPipelining: Bool = true,
-        idleTimeoutConfiguration: IdleStateHandlerConfiguration? = nil,
         threadPoolSize: Int = 2,
         logLevel: Logger.Level? = nil,
         noHTTPServer: Bool = false,
@@ -178,9 +141,6 @@ public struct HBApplicationConfiguration: Sendable {
         self.maxStreamingBufferSize = maxStreamingBufferSize
         self.backlog = 256 // not used by Network framework
         self.reuseAddress = reuseAddress
-        self.tcpNoDelay = true // not used by Network framework
-        self.enableHttpPipelining = enableHttpPipelining
-        self.idleTimeoutConfiguration = idleTimeoutConfiguration
         self.tlsOptions = tlsOptions
 
         self.threadPoolSize = threadPoolSize
@@ -205,9 +165,6 @@ public struct HBApplicationConfiguration: Sendable {
         maxStreamingBufferSize: Int? = nil,
         backlog: Int? = nil,
         reuseAddress: Bool? = nil,
-        tcpNoDelay: Bool? = nil,
-        enableHttpPipelining: Bool? = nil,
-        idleTimeoutConfiguration: IdleStateHandlerConfiguration? = nil,
         threadPoolSize: Int? = nil,
         logLevel: Logger.Level? = nil
     ) -> Self {
@@ -218,9 +175,6 @@ public struct HBApplicationConfiguration: Sendable {
             maxStreamingBufferSize: maxStreamingBufferSize ?? self.maxStreamingBufferSize,
             backlog: backlog ?? self.backlog,
             reuseAddress: reuseAddress ?? self.reuseAddress,
-            tcpNoDelay: tcpNoDelay ?? self.tcpNoDelay,
-            enableHttpPipelining: enableHttpPipelining ?? self.enableHttpPipelining,
-            idleTimeoutConfiguration: idleTimeoutConfiguration ?? self.idleTimeoutConfiguration,
             threadPoolSize: threadPoolSize ?? self.threadPoolSize,
             logLevel: logLevel ?? self.logLevel
         )
@@ -228,28 +182,24 @@ public struct HBApplicationConfiguration: Sendable {
 
     /// return HTTP server configuration
     #if canImport(Network)
-    var httpServer: HBHTTPServer.Configuration {
+    var httpServer: HBServerConfiguration {
         return .init(
             address: self.address,
             serverName: self.serverName,
-            maxUploadSize: self.maxStreamedUploadSize, // we pass down the max streamed upload size here as server assumes everything is streamed
             maxStreamingBufferSize: self.maxStreamingBufferSize,
+            backlog: self.backlog,
             reuseAddress: self.reuseAddress,
-            withPipeliningAssistance: self.enableHttpPipelining,
             tlsOptions: self.tlsOptions
         )
     }
     #else
-    var httpServer: HBHTTPServer.Configuration {
+    var httpServer: HBServerConfiguration {
         return .init(
             address: self.address,
             serverName: self.serverName,
-            maxUploadSize: self.maxStreamedUploadSize, // we pass down the max streamed upload size here as server assumes everything is streamed
             maxStreamingBufferSize: self.maxStreamingBufferSize,
             backlog: self.backlog,
-            reuseAddress: self.reuseAddress,
-            tcpNoDelay: self.tcpNoDelay,
-            withPipeliningAssistance: self.enableHttpPipelining
+            reuseAddress: self.reuseAddress
         )
     }
     #endif
