@@ -26,7 +26,7 @@ extension URLEncodedFormEncoder: HBResponseEncoder {
         return HBResponse(
             status: .ok,
             headers: ["content-type": "application/x-www-form-urlencoded"],
-            body: .byteBuffer(buffer)
+            body: .init(byteBuffer: buffer)
         )
     }
 }
@@ -37,7 +37,7 @@ extension URLEncodedFormDecoder: HBRequestDecoder {
     ///   - type: Type to decode
     ///   - request: Request to decode from
     public func decode<T: Decodable>(_ type: T.Type, from request: HBRequest, context: HBRequestContext) throws -> T {
-        guard var buffer = request.body.buffer,
+        guard case .byteBuffer(var buffer) = request.body,
               let string = buffer.readString(length: buffer.readableBytes)
         else {
             throw HBHTTPError(.badRequest)
