@@ -33,13 +33,13 @@ class TransportServicesTests: XCTestCase {
 
     func testConnect() async throws {
         let eventLoopGroup = NIOTSEventLoopGroup()
-        let server = HBHTTPServer(
-            group: eventLoopGroup,
+        defer { try? eventLoopGroup.syncShutdownGracefully() }
+        try await testServer(
+            childChannelSetup: HTTP1Channel(responder: helloResponder),
             configuration: .init(address: .hostname(port: 0)),
-            responder: HelloResponder(),
+            eventLoopGroup: eventLoopGroup,
             logger: Logger(label: "HB")
-        )
-        try await testServer(server) { client in
+        ) { client in
             let response = try await client.get("/")
             var body = try XCTUnwrap(response.body)
             XCTAssertEqual(body.readString(length: body.readableBytes), "Hello")
@@ -52,19 +52,11 @@ class TransportServicesTests: XCTestCase {
         let tlsOptions = try XCTUnwrap(TSTLSOptions.options(
             serverIdentity: .p12(filename: p12Path, password: "MyPassword")
         ))
-        let configuration = HBHTTPServer.Configuration(
-            address: .hostname(port: 0),
-            serverName: testServerName,
-            tlsOptions: tlsOptions
-        )
-        let server = HBHTTPServer(
-            group: eventLoopGroup,
-            configuration: configuration,
-            responder: HelloResponder(),
-            logger: Logger(label: "HB")
-        )
         try await testServer(
-            server,
+            childChannelSetup: HTTP1Channel(responder: helloResponder),
+            configuration: .init(address: .hostname(port: 0), serverName: testServerName, tlsOptions: tlsOptions),
+            eventLoopGroup: eventLoopGroup,
+            logger: Logger(label: "HB"),
             clientConfiguration: .init(tlsConfiguration: self.getClientTLSConfiguration(), serverName: testServerName)
         ) { client in
             let response = try await client.get("/")
